Add views/reqs validation

+++ flaskr/requests/statuses.py
@@ -1,3 +1,5 @@
+from cerberus import Validator
+
 from flaskr import db
 from flaskr.models.status import Status
 from flaskr.models.lead import Lead
@@ -5,12 +7,20 @@ from flaskr.models.lead import Lead
 
 # Create status
 def create_status(params, request_data):
+    vld = Validator({
+        'name': {'type': 'string', 'required': True},
+        'color': {'type': 'string', 'required': True, 'allowed': ['red', 'pink', 'purple', 'blue', 'green', 'orange']}
+    })
+    is_valid = vld.validate(params)
+    if not is_valid:
+        return {'res': 'err', 'message': 'Invalid params', 'errors': vld.errors}
+
     new_status = Status()
     new_status.veokit_installation_id = request_data['installation_id']
-    new_status.name = params.get('name')
-    new_status.color = params.get('color')
-    new_status.index = Status.query\
-                           .filter_by(veokit_installation_id = request_data['installation_id'])\
+    new_status.name = params['name']
+    new_status.color = params['color']
+    new_status.index = Status.query \
+                           .filter_by(veokit_installation_id=request_data['installation_id']) \
                            .count() - 1
 
     db.session.add(new_status)
@@ -24,8 +34,17 @@ def create_status(params, request_data):
 
 # Update status
 def update_status(params, request_data):
-    status = Status.query\
-        .filter_by(id=params['id'])\
+    vld = Validator({
+        'id': {'type': 'number', 'required': True},
+        'name': {'type': 'string', 'required': True},
+        'color': {'type': 'string', 'required': True, 'allowed': ['red', 'pink', 'purple', 'blue', 'green', 'orange']}
+    })
+    is_valid = vld.validate(params)
+    if not is_valid:
+        return {'res': 'err', 'message': 'Invalid params', 'errors': vld.errors}
+
+    status = Status.query \
+        .filter_by(id=params['id']) \
         .first()
 
     status.name = params.get('name')
@@ -40,9 +59,17 @@ def update_status(params, request_data):
 
 # Update status index
 def update_status_index(params, request_data):
-    statuses = Status.query\
-        .filter_by(veokit_installation_id=request_data['installation_id'])\
-        .order_by(Status.index.asc())\
+    vld = Validator({
+        'id': {'type': 'number', 'required': True},
+        'newIndex': {'type': 'number', 'required': True}
+    })
+    is_valid = vld.validate(params)
+    if not is_valid:
+        return {'res': 'err', 'message': 'Invalid params', 'errors': vld.errors}
+
+    statuses = Status.query \
+        .filter_by(veokit_installation_id=request_data['installation_id']) \
+        .order_by(Status.index.asc()) \
         .all()
 
     status_current_index = next((i for i, s in enumerate(statuses) if s.id == params['id']), None)
@@ -62,15 +89,23 @@ def update_status_index(params, request_data):
 
 # Delete status
 def delete_status(params, request_data):
+    vld = Validator({
+        'id': {'type': 'number', 'required': True},
+        'assignedStatusId': {'type': 'number', 'required': True, 'nullable': True}
+    })
+    is_valid = vld.validate(params)
+    if not is_valid:
+        return {'res': 'err', 'message': 'Invalid params', 'errors': vld.errors}
+
     if params.get('assignedStatusId'):
         Lead.query \
             .filter_by(veokit_installation_id=request_data['installation_id'],
                        status_id=params['id']) \
             .update({'status_id': params['assignedStatusId']})
     else:
-        Lead.query\
+        Lead.query \
             .filter_by(veokit_installation_id=request_data['installation_id'],
-                       status_id=params['id'])\
+                       status_id=params['id']) \
             .delete()
 
     Status.query \
